refactor(#3388): Extend test to check if measurement is removed from asset links
