+++ ui/cypress/support/utils/datalake/DataLakeUtils.ts
@@ -363,10 +363,6 @@ export class DataLakeUtils {
         cy.dataCy('data-explorer-select-data-set-create-btn').click();
     }
 
-    public static goToDatalakeConfiguration() {
-        cy.visit('#/configuration/datalake');
-    }
-
     public static checkResults(
         dataLakeIndex: string,
         fileRoute: string,
@@ -445,23 +441,6 @@ export class DataLakeUtils {
         currentDate.setMonth(currentDate.getMonth() + 1);
 
         return currentDate;
-    }
-
-    public static waitForCountingResults() {
-        cy.dataCy('datalake-number-of-events-spinner', {
-            timeout: 10000,
-        }).should('exist');
-        cy.dataCy('datalake-number-of-events-spinner', {
-            timeout: 10000,
-        }).should('not.exist');
-    }
-
-    public static getDatalakeNumberOfEvents(): Cypress.Chainable<string> {
-        return cy
-            .dataCy('datalake-number-of-events', { timeout: 10000 })
-            .should('be.visible')
-            .invoke('text')
-            .then(text => text.trim());
     }
 
     public static checkRowsDashboardTable(amount: number) {
